requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
albucore==0.0.23
albumentations==1.4.24
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
blis==1.3.0
braceexpand==0.1.7
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
dataclasses==0.6
datasets==4.2.0
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
easydict==1.13
einops==0.8.1
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
executing==2.2.1
ExifRead-nocycle==3.0.1
fastjsonschema==2.21.2
filelock==3.20.0
fire==0.7.1
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
img2dataset==1.47.0
ipykernel==7.0.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lark==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
murmurhash==1.0.13
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
open_clip_torch
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
preshed==3.0.10
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycocotools==2.0.10
pycparser==2.23
pydantic==2.12.1
pydantic_core==2.41.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.27.1
safetensors==0.6.2
scipy==1.16.2
Send2Trash==1.8.3
sentencepiece==0.2.1
sentry-sdk==2.41.0
setuptools==80.9.0
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
standard-imghdr==3.13.0
stringzilla==4.2.1
sympy==1.14.0
termcolor==3.1.0
terminado==0.18.1
thinc==8.3.6
timm==0.9.16
tinycss2==1.4.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
typer==0.19.2
types-python-dateutil==2.9.0.20251008
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wandb==0.22.2
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.1
webcolors==24.11.1
webdataset==0.2.111
webencodings==0.5.1
websocket-client==1.9.0
wheel==0.45.1
wrapt==1.17.3
xmltodict==1.0.2
xxhash==3.6.0
yarl==1.22.0
